refactored transport tests
